+++ tests/requirements/requirements.py
@@ -1,6 +1,6 @@
 # These requirements were auto generated
 # from software requirements specification (SRS)
-# document by TestFlows v2.0.240111.1210833.
+# document by TestFlows v2.0.231215.1221232.
 # Do not edit by hand but re-generate instead
 # using 'tfs requirements generate' command.
 from testflows.core import Specification
@@ -295,10 +295,22 @@ RQ_SRS_026_ClickHouseOperator_Secrets = Requirement(
         '  configuration:\n'
         '    users:\n'
         '      default/password: default\n'
-        '      # reference to namespace/name/field in the secret containing passwords\n'
-        '      user1/k8s_secret_password: test/test-011-secret/pwduser1\n'
-        '      user2/k8s_secret_password_sha256_hex: test/test-011-secret/pwduser2\n'
-        '      user3/k8s_secret_password_double_sha1_hex: test/test-011-secret/pwduser3\n'
+        "      # reference a secret in the CHI's own namespace containing passwords\n"
+        '      user1/password:\n'
+        '        valueFrom:\n'
+        '          secretKeyRef:\n'
+        '            name: test-011-secret\n'
+        '            key: pwduser1\n'
+        '      user2/password_sha256_hex:\n'
+        '        valueFrom:\n'
+        '          secretKeyRef:\n'
+        '            name: test-011-secret\n'
+        '            key: pwduser2\n'
+        '      user3/password_double_sha1_hex:\n'
+        '        valueFrom:\n'
+        '          secretKeyRef:\n'
+        '            name: test-011-secret\n'
+        '            key: pwduser3\n'
         '    clusters:\n'
         '    - name: default\n'
         '      layout:\n'
@@ -8397,10 +8409,22 @@ spec:
   configuration:
     users:
       default/password: default
-      # reference to namespace/name/field in the secret containing passwords
-      user1/k8s_secret_password: test/test-011-secret/pwduser1
-      user2/k8s_secret_password_sha256_hex: test/test-011-secret/pwduser2
-      user3/k8s_secret_password_double_sha1_hex: test/test-011-secret/pwduser3
+      # reference a secret in the CHI's own namespace containing passwords
+      user1/password:
+        valueFrom:
+          secretKeyRef:
+            name: test-011-secret
+            key: pwduser1
+      user2/password_sha256_hex:
+        valueFrom:
+          secretKeyRef:
+            name: test-011-secret
+            key: pwduser2
+      user3/password_double_sha1_hex:
+        valueFrom:
+          secretKeyRef:
+            name: test-011-secret
+            key: pwduser3
     clusters:
     - name: default
       layout: